IMS-VT: Do not toggle full screen mode when not in video call.
If far end view is clicked almost same time when video call is
downgraded from remote end, there is a chance that click event
is handled after exiting video mode. This will cause call controls
to be removed from UI. To fix this, ignore toggling of fullscreen
mode when not in video mode.

Change-Id: I08c773d9bcc3c99b9aee3b2bb235f3d38762d380
CRs-Fixed: 1031498