IMS-VT: Fix orientation change not working after back key press
When InCall activity is recreated after finishing by back key press,
there is a chance that VideoCallPresenter tries to enable
orientation change even before activity is registered with
IncallPresenter. In such cases the cached orientation mode is
updated in VideoCallPresenter even though it was not really updated
in InCallActivity. To fix this, check if new orientation mode was
actually updated before updating cached value in VideoCallPresenter.

Change-Id: Iedc06167bf2edd50c7405a09cd99e68fdc3f4f7c
Crs-Fixed: 1050432